ECS - Refactored constants to `GameConstants`, updated `NetworkSyncComponent` to use floats, and implemented directional rotation in `MovementSystem`.

+++ core/src/main/java/com/fatpiggies/game/model/ecs/components/NetworkSyncComponent.java
@@ -3,6 +3,6 @@ package com.fatpiggies.game.model.ecs.components;
 import com.badlogic.ashley.core.Component;
 
 public class NetworkSyncComponent implements Component {
-    public double targetX;
-    public double targetY;
+    public float targetX;
+    public float targetY;
 }
